+++ Source/Surface.hpp
@@ -73,7 +73,7 @@ public:
 					cSurface( const cDimension& pDimension );
 					~cSurface();
 
-	void			clearBuffer(uint8 pColor = 0);
+	void			clearBuffer(size_t pColor = 0);
 
 	void			draw();						// Draw image to SDL Surface
     void            mergeSurfaceBuffer(const cSurface* pFrom);
